chore: revise file strucutre
